make it stop asking for info forever
